+++ consensus/src/utils.rs
@@ -65,74 +65,6 @@ impl<TYPES: NodeType, LEAF: LeafType<NodeType = TYPES>> Deref for View<TYPES, LE
     }
 }
 
-/// Alias for the [`ProcessedConsensusMessage`] type of a [`NodeImplementation`].
-type ProcessedConsensusMessageType<TYPES, I> = <<I as NodeImplementation<TYPES>>::ConsensusMessage as ConsensusMessageType<TYPES, I>>::ProcessedConsensusMessage;
-
-/// struct containing messages for a view to send to a replica or DA committee member.
-#[derive(Clone)]
-pub struct ViewQueue<TYPES: NodeType, I: NodeImplementation<TYPES>> {
-    /// to send networking events to a replica or DA committee member.
-    pub sender_chan: UnboundedSender<ProcessedConsensusMessageType<TYPES, I>>,
-
-    /// to recv networking events for a replica or DA committee member.
-    pub receiver_chan: Arc<Mutex<UnboundedReceiver<ProcessedConsensusMessageType<TYPES, I>>>>,
-
-    /// `true` if this queue has already received a proposal
-    pub has_received_proposal: Arc<AtomicBool>,
-}
-
-impl<TYPES: NodeType, I: NodeImplementation<TYPES>> Default for ViewQueue<TYPES, I> {
-    /// create new view queue
-    fn default() -> Self {
-        let (s, r) = unbounded();
-        ViewQueue {
-            sender_chan: s,
-            receiver_chan: Arc::new(Mutex::new(r)),
-            has_received_proposal: Arc::new(AtomicBool::new(false)),
-        }
-    }
-}
-
-/// metadata for sending information to replica (and in the future, the leader)
-pub struct SendToTasks<TYPES: NodeType, I: NodeImplementation<TYPES>> {
-    /// the current view number
-    /// this should always be in sync with `Consensus`
-    pub cur_view: TYPES::Time,
-
-    /// a map from view number to ViewQueue
-    /// one of (replica|next leader)'s' task for view i will be listening on the channel in here
-    pub channel_map: BTreeMap<TYPES::Time, ViewQueue<TYPES, I>>,
-}
-
-impl<TYPES: NodeType, I: NodeImplementation<TYPES>> SendToTasks<TYPES, I> {
-    /// create new sendtosasks
-    #[must_use]
-    pub fn new(view_num: TYPES::Time) -> Self {
-        SendToTasks {
-            cur_view: view_num,
-            channel_map: BTreeMap::default(),
-        }
-    }
-}
-
-/// Channels for sending/recv-ing proposals and votes.
-pub struct ChannelMaps<TYPES: NodeType, I: NodeImplementation<TYPES>> {
-    /// Channel for the next consensus leader or DA leader.
-    proposal_channel:Arc<RwLock<SendToTasks<TYPES, I>>>,
-
-    /// Channel for the replica or DA committee member.
-    vote_channel:Arc<RwLock<SendToTasks<TYPES, I>>>,
-  }
-
-impl<TYPES: NodeType, I: NodeImplementation<TYPES>> ChannelMaps<TYPES, I> {
-    pub fn new(start_view: TYPES::Time) -> Self {
-        Self {
-            proposal_channel:Arc::new(RwLock::new(SendToTasks::new(start_view))),
-            vote_channel: Arc::new(RwLock::new(SendToTasks::new(start_view)))
-        }
-    }
-}
-
 /// This exists so we can perform state transitions mutably
 #[derive(Debug)]
 pub struct View<TYPES: NodeType, LEAF: LeafType<NodeType = TYPES>> {
